# src/lib/api/openapi.yaml
openapi: 3.1.0
info:
  title: Shelly Monitor API
  description: |
    Enterprise-grade API for monitoring and controlling Shelly IoT devices.
    
    ## Authentication
    All endpoints except health checks require authentication via JWT bearer token.
    
    ## Rate Limiting
    - Global: 1000 requests per 15 minutes
    - Auth endpoints: 5 requests per 15 minutes
    - API endpoints: 100 requests per minute
    - Control endpoints: 20 requests per minute
  version: 1.0.0
  contact:
    name: API Support
    email: api@shellymonitor.com
    url: https://github.com/blipee-dev/shelly-monitor
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.shellymonitor.com/v1
    description: Production server
  - url: https://staging-api.shellymonitor.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticate user and receive JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/register:
    post:
      tags:
        - Auth
      summary: User registration
      description: Create new user account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh token
      description: Get new access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Refresh token
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices:
    get:
      tags:
        - Devices
      summary: List devices
      description: Get all devices for authenticated user
      operationId: listDevices
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          description: Filter by device type
          schema:
            type: string
            enum: [plus_2pm, motion_2]
        - name: status
          in: query
          description: Filter by online status
          schema:
            type: string
            enum: [online, offline]
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Devices
      summary: Add device
      description: Add new device to monitoring
      operationId: addDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDeviceRequest'
      responses:
        '201':
          description: Device added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /devices/{deviceId}/control:
    post:
      tags:
        - Devices
      summary: Control device
      description: Send control command to device (Plus 2PM only)
      operationId: controlDevice
      parameters:
        - $ref: '#/components/parameters/DeviceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlRequest'
      responses:
        '200':
          description: Command executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  result:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  parameters:
    DeviceIdParam:
      name: deviceId
      in: path
      required: true
      description: Device ID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        name:
          type: string
          minLength: 2
          maxLength: 100
    AuthResponse:
      type: object
      required:
        - user
        - accessToken
        - refreshToken
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Token expiry in seconds
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, user, viewer]
        createdAt:
          type: string
          format: date-time
    Device:
      type: object
      required:
        - id
        - name
        - type
        - ipAddress
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [plus_2pm, motion_2]
        model:
          type: string
        ipAddress:
          type: string
          format: ipv4
        macAddress:
          type: string
          pattern: '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
        firmwareVersion:
          type: string
        authEnabled:
          type: boolean
        settings:
          type: object
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AddDeviceRequest:
      type: object
      required:
        - name
        - type
        - ipAddress
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          type: string
          enum: [plus_2pm, motion_2]
        ipAddress:
          type: string
          format: ipv4
        authEnabled:
          type: boolean
        authUsername:
          type: string
        authPassword:
          type: string
...
    ControlRequest:
      type: object
      required:
        - action
        - channel
        - state
      properties:
        action:
          type: string
          enum: [switch]
        channel:
          type: integer
          minimum: 0
          maximum: 1
        state:
          type: boolean
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INVALID_REQUEST
              message: Invalid request parameters
              timestamp: "2024-01-01T00:00:00.000Z"
              requestId: "req_123456"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
              timestamp: "2024-01-01T00:00:00.000Z"
              requestId: "req_123456"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
              timestamp: "2024-01-01T00:00:00.000Z"
              requestId: "req_123456"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: CONFLICT
              message: Resource already exists
              timestamp: "2024-01-01T00:00:00.000Z"
              requestId: "req_123456"
    RateLimited:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RATE_LIMITED
              message: Rate limit exceeded
              details:
                retryAfter: 60
              timestamp: "2024-01-01T00:00:00.000Z"
              requestId: "req_123456"
    ServiceUnavailable:
      description: Service Unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: SERVICE_UNAVAILABLE
              message: Service temporarily unavailable
              timestamp: "2024-01-01T00:00:00.000Z"
              requestId: "req_123456"